GCStress: Don't force Cooperative mode
This change removes the logic in GC-Coverage implementation
that puts a thread in premptive mode to cooperative mode
in order to do a GCStress collection.

With this change, GCStress is only performed when we encounter
a trap for GC in cooperative mode. When the trap is encountered
in premptive mode, the GC-Trap is replaced with the original
instruction without starting a collection at that point.
This is in line with the expectation that native threads typically
don't start garbage-collection.

Artificially moving the thread to cooperative mode for forcing GCStress
from a thread executing in Premptive mode will lead to complexity
in GCStress setup and ThreadSuspension logic.