Implement GC root scanning for Hijack-frames
This change implements GCScanRoots() method for Hijacke-frames
based on the ReturnKind information available from the GcInfo.
If the exact ReturnKind is not available in the GcInfo, the
thread-suspension logic will compute the ReturnKind based on
the method-signature.

As a result of this change, several hijack-helpers in the VM
are cleaned up. There's only one implementation of HijackWorker()
which handles all returnKinds, and one assembly helper for handling
all methods that return a GC value.

This change also fixes NativeImageDumper::DumpReadyToRunMethod()
to use the correct GcInfo version based on the ReadyToRun header.